app/requirements.dev.txt
django-stubs
python-dotenv